Added login views and amended add/del views and index query.

+++ views.py
@@ -1,13 +1,14 @@
 from django.http import HttpResponse
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, reverse
 from django.contrib.auth import authenticate,login,logout
 from django.contrib.auth.decorators import login_required
+from django.core.exceptions import ObjectDoesNotExist
 from .models import tracking_numbers
 import re
 
-@login_required(login_url='/admin/login')
+@login_required(login_url='/login')
 def index(request):
-	refs = tracking_numbers.objects.values()
+	refs = tracking_numbers.objects.values().filter(user_id=request.user.id)
 	csv_refs = ''
 	csv_list = []
 	for i in refs:
@@ -18,14 +19,35 @@ def index(request):
 	return render(request, 'index.html', context)
 
 def delete(request, id_del):
-	ostju = tracking_numbers.objects.get(id=id_del)
-	ostju.delete()
-	return redirect('/')
-#	return HttpResponse('<script>javascript:alert("Not gonna happen");window.location.assign("/");</script>')
+	try:
+		ostju = tracking_numbers.objects.get(id=id_del, user_id=request.user.id)
+	except ObjectDoesNotExist as e:
+		ostju = ''
+	if ostju:
+		ostju.delete()
+		return redirect('/')
+	else:
+		return HttpResponse('<script>javascript:alert("Invalid request.");window.location.assign("/");</script>')
 
 def add(request):
+	user_id = request.user.id
 	tracking_number = request.POST.get('tracking_number')
 	description = request.POST.get('description')
-	new_data  =  tracking_numbers(tracking_number=tracking_number, description=description)
+	new_data  =  tracking_numbers(tracking_number=tracking_number, description=description, user_id=user_id)
 	new_data.save()
 	return redirect('/')
+
+def loginF(request):
+	return render(request, 'login.html')
+
+def loginP(request):
+	username = request.POST.get('username')
+	password = request.POST.get('password')
+	user = authenticate(username=username, password=password)
+	if user:
+		login(request, user)
+		return redirect('/')
+	else:
+		error_message = "Invalid credentials."
+		context = { error_message: 'blablablahhh' }
+		return redirect('/', error_message)
